refactor: shared/ui/index.ts 분리하기
