+++ src/shared/ui/Card/CardContent.tsx
@@ -1,0 +1,7 @@
+import { forwardRef } from "react";
+import { DivProps } from "./card.type";
+
+export const CardContent = forwardRef<HTMLDivElement, DivProps>(({ className = "", ...props }, ref) => (
+  <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
+));
+CardContent.displayName = "CardContent";
