+++ src/shared/ui/Card/card.type.ts
@@ -1,0 +1,3 @@
+export interface DivProps extends React.HTMLAttributes<HTMLDivElement> {
+  className?: string;
+}
