+++ src/shared/ui/Dialog/DialogTrigger.tsx
@@ -1,0 +1,3 @@
+import * as DialogPrimitive from "@radix-ui/react-dialog";
+
+export const DialogTrigger = DialogPrimitive.Trigger;
