+++ src/shared/ui/Dialog/dialog.index.ts
@@ -1,0 +1,7 @@
+export { Dialog } from "./Dialog";
+export { DialogContent } from "./DialogContent";
+export { DialogHeader } from "./DialogHeader";
+export { DialogOverlay } from "./DialogOverlay";
+export { DialogPortal } from "./DialogPortal";
+export { DialogTitle } from "./DialogTitle";
+export { DialogTrigger } from "./DialogTrigger";
