+++ src/shared/ui/Select/select.index.ts
@@ -1,0 +1,6 @@
+export { SelectContent } from "./SelectContent";
+export { SelectItem } from "./SelectItem";
+export { SelectTrigger } from "./SelectTrigger";
+export { Select } from "./Select";
+export { SelectGroup } from "./SelectGroup";
+export { SelectValue } from "./SelectValue";
