+++ src/shared/ui/Table/TableBody.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from "react";
+
+export const TableBody = forwardRef<HTMLTableSectionElement, React.HTMLAttributes<HTMLTableSectionElement>>(
+  ({ className = "", ...props }, ref) => (
+    <tbody ref={ref} className={`[&_tr:last-child]:border-0 ${className}`} {...props} />
+  ),
+);
+TableBody.displayName = "TableBody";
