+++ src/shared/ui/Table/TableHeader.tsx
@@ -1,0 +1,6 @@
+import { forwardRef } from "react";
+
+export const TableHeader = forwardRef<HTMLTableSectionElement, React.HTMLAttributes<HTMLTableSectionElement>>(
+  ({ className = "", ...props }, ref) => <thead ref={ref} className={`[&_tr]:border-b ${className}`} {...props} />,
+);
+TableHeader.displayName = "TableHeader";
