+++ src/shared/ui/Table/table.index.ts
@@ -1,0 +1,6 @@
+export { Table } from "./Table";
+export { TableBody } from "./TableBody";
+export { TableCell } from "./TableCell";
+export { TableHead } from "./TableHead";
+export { TableHeader } from "./TableHeader";
+export { TableRow } from "./TableRow";
