+++ src/shared/ui/index.ts
@@ -1,0 +1,15 @@
+export { Button } from "./Button";
+export { Card, CardContent, CardHeader, CardTitle } from "./Card/card.index";
+export {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogOverlay,
+  DialogPortal,
+  DialogTitle,
+  DialogTrigger,
+} from "./Dialog/dialog.index";
+export { Input } from "./Input";
+export { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from "./Select/select.index";
+export { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "./Table/table.index";
+export { Textarea } from "./Textarea";
